+++ src/main/java/com/docshifter/core/metrics/entities/DocumentCounter.java
@@ -6,7 +6,6 @@ import lombok.Setter;
 
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import java.math.BigInteger;
 
 
 @Entity(name="DocumentCounter")
@@ -18,14 +17,6 @@ public class DocumentCounter {
 @Id
 private String task_id;
 
-private BigInteger counts;
-
-//public void setId(String id) { this.task_id = id;}
-//
-//public String getId() { return this.task_id;}
-//
-//public void setCounts(int counts) { this.counts=counts;}
-//
-//public int getCounts() { return this.counts;}
+private long counts;
 
 }
