+++ src/main/java/com/docshifter/core/metrics/repositories/DocumentCounterRepository.java
@@ -7,9 +7,9 @@ import org.springframework.data.jpa.repository.Query;
 public interface DocumentCounterRepository extends JpaRepository<DocumentCounter, String> {
 
     @Query("SELECT SUM(c.counts) FROM DocumentCounter c")
-    int selectTotalCounts();
+    long selectTotalCounts();
 
     @Query("SELECT COUNT(c.task_id) FROM DocumentCounter c")
-    int selectSuccessfulWorkflows();
+    long selectSuccessfulWorkflows();
 
 }
